ignore test.

+++ modules/historique/clear_prev_history_files.py
@@ -4,4 +4,5 @@ def clear_prev_history_files():
   history_files = os.listdir('files/historique')
 
   for history_file in history_files:
-    os.remove(os.path.join('files/historique', history_file))
+    if (history_file != '.gitignore'):
+       os.remove(os.path.join('files/historique', history_file))
