+++ modules/ordonnance/remove_old_ord_files.py
@@ -3,4 +3,5 @@ import os
 
 def remove_old_ord_files():
   for ordo in os.listdir('files/ordonnance'):
-    os.remove(os.path.join('files/ordonnance', ordo))
+    if (ordo != '.gitignore'):
+      os.remove(os.path.join('files/ordonnance', ordo))
